mv ChatNavList -> ThreadsList
